backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.1
langchain-community==0.3.1
langchain-groq==0.2.0
groq==0.11.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
python-multipart==0.0.12
httpx==0.27.2
